strategy/requirements.txt
asyncpg==0.29.0
nats-py==2.7.2
uvloop==0.19.0
orjson==3.9.15
PyYAML==6.0.2
